move book component to file
