+++ front/src/components/book/BookGrid.tsx
@@ -1,20 +1,21 @@
 import { Grid, Text } from "@chakra-ui/react"
 import axios from "axios"
 import { useEffect, useState } from "react"
-import { API_ENDPOINT } from "../utils/apiEndPoint"
+import { API_ENDPOINT } from "../../utils/apiEndPoint"
 import { BookItem } from "./BookItem"
 
 export const BookGrid = () => {
   const [bookData, setBookData] = useState([])
 
   useEffect(() => {
-    axios.get(`${API_ENDPOINT}/books`)
-    .then((res) => {
-      setBookData(res.data.books)
-    })
-    .catch((err) => {
-      console.log(err)
-    })
+    axios
+      .get(`${API_ENDPOINT}/books`)
+      .then((res) => {
+        setBookData(res.data.books)
+      })
+      .catch((err) => {
+        console.log(err)
+      })
   }, [])
 
   return (
@@ -29,15 +30,14 @@ export const BookGrid = () => {
           {bookData.map((data) => (
             <BookItem
               bookId={data.id}
-              title={data.title}
-              imageUrl={data.epub_url}
+              title={data.Title}
+              imageUrl={data.test1}
             />
           ))}
         </>
       ) : (
         <Text>本棚に本がありません</Text>
-      )
-      }
+      )}
     </Grid>
   )
 }
