+++ front/src/components/book/BookItem.tsx
