+++ front/src/pages/book/[id].tsx
@@ -1,7 +1,7 @@
-import { useRouter } from "next/router";
-import { memo, VFC } from "react";
-import { AddBook } from "../../components/AddBook";
-import { getAsStringFromArray } from "../../utils/getAsStringFromArray";
+import { useRouter } from "next/router"
+import { memo, VFC } from "react"
+import { AddBook } from "../../components/book/AddBook"
+import { getAsStringFromArray } from "../../utils/getAsStringFromArray"
 
 const App: VFC = memo(() => {
   const router = useRouter()
