+++ front/src/pages/read/[id].tsx
@@ -3,7 +3,9 @@ import dynamic from "next/dynamic"
 import { useRouter } from "next/router"
 import { getAsStringFromArray } from "../../utils/getAsStringFromArray"
 
-const Reader = dynamic(() => import("../../components/Reader"), { ssr: false })
+const Reader = dynamic(() => import("../../components/book/Reader"), {
+  ssr: false,
+})
 
 const App = () => {
   const router = useRouter()
